can update sensitivity

+++ src/cameras/configurations.py
@@ -3,7 +3,7 @@ CONFIGURATIONS_KEYS = ["sensitivity", "recording"]
 
 
 class Configurations:
-    _sensitivity: int
+    _sensitivity: float
     recording: bool
 
     def __init__(self, **configurations):
@@ -15,7 +15,7 @@ class Configurations:
         return self._sensitivity
 
     @sensitivity.setter
-    def sensitivity(self, sensitivity: int):
+    def sensitivity(self, sensitivity: float):
         if sensitivity < 0 or sensitivity > 1:
             raise Exception("Sensitivity must be a number between 0 and 1")
